don't validate video/audio parameters when copying
These checks happen to validate the capabilities of the ffmpeg binary.
More generally these can perhaps be made parallel since they are a bit
slow. It's also questionable if they really need to be a part of startup
or if these checks can be pushed back later and made during initial
parameter query.

One still may question if these are even needed since many ffmpeg
binaries have the support compiled in even if it's not available and
would fail anyway. This means these checks often won't improve user
experience anyway despite having a noticeable effect on the startup time
due to the overhead of needing to query the external ffmpeg processes.