remove EncodeSession ownership by FileStreamVals
This is only used to find out the probe arguments and whether or not a
particular FileStreamVals instance should allow probing. For now we pass
the StaticEncodeVars in the instance constructor and set these
variables, but it might make sense to separate this further and pass
these to the constructor directly.