convert ffmpeg.num to singledispatch
The base case is just the str, which is what the function is intended
for. The other numeric cases are allowed to the extent this function is
used for json deserializing from ffmpeg output, but it may be worth
considering removing the other uses and making it solely a function of
str.